+++ src/orb/____inner_core/geom/_Rect.java
@@ -6,10 +6,10 @@ package orb.____inner_core.geom;
 // todo? it should be different from AABB cause it does not have to be axis aligned
 public interface _Rect {
 
-    double x1();
-    double y1();
-    double x2();
-    double y2();
+    float x1();
+    float y1();
+    float x2();
+    float y2();
 
     // methods
 }
