+++ src/orb/____inner_core/geom/_Triangle.java
@@ -1,6 +1,6 @@
 package orb.____inner_core.geom;
 
-import nl.doekewartena.orb.inner_core.IC_Math;
+import orb.____inner_core.IC_Math;
 
 /**
  * Created by doekewartena on 5/23/15.
@@ -8,26 +8,26 @@ import nl.doekewartena.orb.inner_core.IC_Math;
 
 public interface _Triangle {
 
-    double x1();
-    double y1();
-    double x2();
-    double y2();
-    double x3();
-    double y3();
+    float x1();
+    float y1();
+    float x2();
+    float y2();
+    float x3();
+    float y3();
 
 
-    default boolean pointInside(double x, double y) {
+    default boolean pointInside(float x, float y) {
         return IC_Math.pointInTriangle(x, y, x1(), y1(), x2(), y2(), x3(), y3());
     }
 
-    default double area() {
+    default float area() {
         return IC_Math.getTriangleArea(x1(), y1(), x2(), y2(), x3(), y3());
     }
 
     default int compareArea(_Triangle t) { // instead of comparing to other triangle, compare to other shape?
 
-        double a1 = area();
-        double a2 = t.area();
+        float a1 = area();
+        float a2 = t.area();
 
         if      (a1 > a2) return 1;
         else if (a1 < a2) return -1;
