+++ src/orb/____inner_core/geom/_Vec4.java
@@ -7,9 +7,9 @@ package orb.____inner_core.geom;
 
 public interface _Vec4<C extends _Vec4> extends _Vec3<C> {
 
-    double w();
+    float w();
 
-    C set(double x, double y, double z, double w);
+    C set(float x, float y, float z, float w);
 
     default C set(_Vec4 o) {
         set(o.x(), o.y(), o.z(), o.w());
@@ -17,7 +17,7 @@ public interface _Vec4<C extends _Vec4> extends _Vec3<C> {
     }
 
 
-    default C add(double x, double y, double z, double w) {
+    default C add(float x, float y, float z, float w) {
         set(x() + x, y() + y, z() + z, w() + w);
         return (C) this;
     }
@@ -27,7 +27,7 @@ public interface _Vec4<C extends _Vec4> extends _Vec3<C> {
         return (C) this;
     }
 
-    default C sub(double x, double y, double z, double w) {
+    default C sub(float x, float y, float z, float w) {
         set(x()-x, y()-y, z()-z, w()-w);
         return (C) this;
     }
@@ -38,53 +38,53 @@ public interface _Vec4<C extends _Vec4> extends _Vec3<C> {
     }
 
 
-    default C mult(double n) {
+    default C mult(float n) {
         set(x()*n, y()*n, z()*n, w()*n);
         return (C) this;
     }
 
-    default C div(double n) {
+    default C div(float n) {
         set(x()/n, y()/n, z()/n, w()/n);
         return (C) this;
     }
 
-    default double mag() {
-        return Math.sqrt(x() * x() + y() * y() + z() * z() + w() * w());
+    default float mag() {
+        return (float)Math.sqrt(x() * x() + y() * y() + z() * z() + w() * w());
     }
 
-    default double magSq() {
+    default float magSq() {
         return x()*x() + y()*y() + z()*z() + w()*w();
     }
 
-    default double dist(double x2, double y2, double z2, double w2) {
-        double dx = x() - x2;
-        double dy = y() - y2;
-        double dz = z() - z2;
-        double dw = w() - w2;
-        return Math.sqrt(dx * dx + dy * dy + dz * dz + dw * dw);
+    default float dist(float x2, float y2, float z2, float w2) {
+        float dx = x() - x2;
+        float dy = y() - y2;
+        float dz = z() - z2;
+        float dw = w() - w2;
+        return (float)Math.sqrt(dx * dx + dy * dy + dz * dz + dw * dw);
     }
 
-    default double dist(_Vec4 o) {
+    default float dist(_Vec4 o) {
         return dist(o.x(), o.y(), o.z(), o.w());
     }
 
-    default double distSq(double x2, double y2, double z2, double w2) {
-        double dx = x() - x2;
-        double dy = y() - y2;
-        double dz = z() - z2;
-        double dw = w() - w2;
+    default float distSq(float x2, float y2, float z2, float w2) {
+        float dx = x() - x2;
+        float dy = y() - y2;
+        float dz = z() - z2;
+        float dw = w() - w2;
         return dx*dx + dy*dy + dz*dz + dw*dw;
     }
 
-    default double distSq(_Vec4 o) {
+    default float distSq(_Vec4 o) {
         return distSq(o.x(), o.y(), o.z(), o.w());
     }
 
-    default double dot(double x2, double y2, double z2, double w2) {
+    default float dot(float x2, float y2, float z2, float w2) {
         return x()*x2 + y()*y2 + z()*z2 + w()*w2;
     }
 
-    default double dot(_Vec4 o) {
+    default float dot(_Vec4 o) {
         return dot(o.x(), o.y(), o.z(), o.w());
     }
