+++ src/orb/____inner_core/util/function/_GetX_T.java
@@ -4,5 +4,5 @@ package orb.____inner_core.util.function;
  * Created by doekewartena on 9/11/15.
  */
 public interface _GetX_T<T> {
-    double getX(T t);
+    float getX(T t);
 }
