+++ src/orb/____inner_core/util/function/unused/_GetY.java
@@ -4,5 +4,5 @@ package orb.____inner_core.util.function.unused;
  * Created by doekewartena on 9/12/15.
  */
 public interface _GetY {
-    double getY();
+    float getY();
 }
